+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # flowery 0.0.1.9000
 
+* Renamed `iterator` class to `flowery_iterator`.
+
 * Exhausted iterators no longer throw when called again. Instead, they
   return `NULL` (#6).